Updated T5 tokenizer to match new Transformers behaviour (addition of EOS if missing at tokenizer level)

+++ main/src/vocab/base_vocab.rs
@@ -382,10 +382,10 @@ mod tests {
         let base_vocab = BaseVocab::from_file(path.to_path_buf().to_str().unwrap())?;
 
         //        When & Then
-        assert_eq!(base_vocab.id_to_token(&(0 as i64)), "hello");
-        assert_eq!(base_vocab.id_to_token(&(1 as i64)), "world");
-        assert_eq!(base_vocab.id_to_token(&(3 as i64)), "!");
-        assert_eq!(base_vocab.id_to_token(&(2 as i64)), "[UNK]");
+        assert_eq!(base_vocab.id_to_token(&(0_i64)), "hello");
+        assert_eq!(base_vocab.id_to_token(&(1_i64)), "world");
+        assert_eq!(base_vocab.id_to_token(&(3_i64)), "!");
+        assert_eq!(base_vocab.id_to_token(&(2_i64)), "[UNK]");
 
         drop(path);
         Ok(())
