+++ main/src/vocab/prophetnet_vocab.rs
@@ -280,15 +280,15 @@ mod tests {
         let bert_vocab = ProphetNetVocab::from_file(path.to_path_buf().to_str().unwrap())?;
 
         //        When & Then
-        assert_eq!(bert_vocab.id_to_token(&(0 as i64)), "hello");
-        assert_eq!(bert_vocab.id_to_token(&(1 as i64)), "world");
-        assert_eq!(bert_vocab.id_to_token(&(3 as i64)), "!");
-        assert_eq!(bert_vocab.id_to_token(&(2 as i64)), "[UNK]");
-        assert_eq!(bert_vocab.id_to_token(&(7 as i64)), "[PAD]");
-        assert_eq!(bert_vocab.id_to_token(&(6 as i64)), "[MASK]");
-        assert_eq!(bert_vocab.id_to_token(&(4 as i64)), "[X_SEP]");
-        assert_eq!(bert_vocab.id_to_token(&(5 as i64)), "[SEP]");
-        assert_eq!(bert_vocab.id_to_token(&(8 as i64)), "[CLS]");
+        assert_eq!(bert_vocab.id_to_token(&(0_i64)), "hello");
+        assert_eq!(bert_vocab.id_to_token(&(1_i64)), "world");
+        assert_eq!(bert_vocab.id_to_token(&(3_i64)), "!");
+        assert_eq!(bert_vocab.id_to_token(&(2_i64)), "[UNK]");
+        assert_eq!(bert_vocab.id_to_token(&(7_i64)), "[PAD]");
+        assert_eq!(bert_vocab.id_to_token(&(6_i64)), "[MASK]");
+        assert_eq!(bert_vocab.id_to_token(&(4_i64)), "[X_SEP]");
+        assert_eq!(bert_vocab.id_to_token(&(5_i64)), "[SEP]");
+        assert_eq!(bert_vocab.id_to_token(&(8_i64)), "[CLS]");
 
         drop(path);
         Ok(())
